feat: Implement ArmadaManagerVaults for deposit and withdrawal transactions
- Added new class `ArmadaManagerVaults` to handle deposit and withdrawal transactions for Armada protocol.
- Implemented methods for creating new deposit transactions, updating deposits, and withdrawing tokens.
- Introduced vault switching functionality with a placeholder for future implementation.
- Enhanced transaction metadata to include vault switch details.
- Updated SDK common types to include `VaultSwitchTransactionInfo` and deprecated `ExtendedTransactionInfo`.
- Added necessary utility methods for previewing deposits, withdrawals, and handling multicall operations.
