+++ sdk/armada-protocol-service/src/common/implementation/ArmadaManagerClaims.ts
@@ -6,6 +6,7 @@ import {
 } from '@summerfi/armada-protocol-abis'
 import {
   type IArmadaManagerClaims,
+  type IArmadaManagerUtils,
   getAllMerkleClaims,
   getDeployedContractAddress,
   getDeployedGovRewardsManagerAddress,
@@ -17,7 +18,6 @@ import {
   type HexData,
   type IAddress,
   type IChainInfo,
-  type IToken,
   type IUser,
 } from '@summerfi/sdk-common'
 import { encodeFunctionData } from 'viem'
@@ -34,7 +34,7 @@ export class ArmadaManagerClaims implements IArmadaManagerClaims {
   private _blockchainClientProvider: IBlockchainClientProvider
   private _contractsProvider: IContractsProvider
   private _configProvider: IConfigurationProvider
-  private _getSummerToken: (params: { chainInfo: IChainInfo }) => IToken
+  private _utils: IArmadaManagerUtils
 
   private _supportedChains: IChainInfo[]
   private _hubChainInfo: IChainInfo
@@ -50,7 +50,7 @@ export class ArmadaManagerClaims implements IArmadaManagerClaims {
     supportedChains: IChainInfo[]
     hubChainInfo: IChainInfo
     rewardsRedeemerAddress: IAddress
-    getSummerToken: (params: { chainInfo: IChainInfo }) => IToken
+    utils: IArmadaManagerUtils
     subgraphManager: IArmadaSubgraphManager
   }) {
     this._blockchainClientProvider = params.blockchainClientProvider
@@ -59,7 +59,7 @@ export class ArmadaManagerClaims implements IArmadaManagerClaims {
     this._supportedChains = params.supportedChains
     this._hubChainInfo = params.hubChainInfo
     this._rewardsRedeemerAddress = params.rewardsRedeemerAddress
-    this._getSummerToken = params.getSummerToken
+    this._utils = params.utils
     this._subgraphManager = params.subgraphManager
     const _distributionsBaseUrl = this._configProvider.getConfigurationItem({
       name: 'SDK_DISTRIBUTIONS_BASE_URL',
@@ -175,7 +175,7 @@ export class ArmadaManagerClaims implements IArmadaManagerClaims {
     })
 
     const rewardsManagerAddress = getDeployedGovRewardsManagerAddress()
-    const rewardToken = this._getSummerToken({ chainInfo: this._hubChainInfo })
+    const rewardToken = this._utils.getSummerToken({ chainInfo: this._hubChainInfo })
 
     return client.readContract({
       abi: GovernanceRewardsManagerAbi,
