+++ sdk/sdk-common/src/orders/common/types/DEPRECATED.ts
@@ -1,0 +1,14 @@
+import type {
+  ApproveTransactionInfo,
+  DepositTransactionInfo,
+  WithdrawTransactionInfo,
+} from './ExtendedTransactionInfo'
+
+/**
+ * @interface ExtendedTransactionInfo
+ * @deprecated DONT TOUCH THIS!!!
+ */
+export type ExtendedTransactionInfo =
+  | ApproveTransactionInfo
+  | DepositTransactionInfo
+  | WithdrawTransactionInfo
